small fixes to temp message encoding end decoding

+++ ehm-common/messagetypes.h
@@ -95,8 +95,8 @@ typedef PACKEDSTRUCT {
 } AccelerometerMessage;
 
 typedef PACKEDSTRUCT {
-	double Vobj;
-	double Tenv;
+	int Vobj;
+	int Tenv;
 } RawTemperatureMessage;
 
 typedef PACKEDSTRUCT {
